ARROW-15282: [C++][FlightRPC] Refactor out CUDA tests
