+++ cpp/src/arrow/flight/CMakeLists.txt
@@ -282,14 +282,6 @@ if(ARROW_BUILD_TESTS OR ARROW_BUILD_BENCHMARKS)
   add_dependencies(arrow_flight flight-test-server)
 endif()
 
-if(ARROW_CUDA AND ARROW_BUILD_TESTS)
-  add_arrow_test(flight_cuda_test
-                 STATIC_LINK_LIBS
-                 ${ARROW_FLIGHT_TEST_LINK_LIBS}
-                 LABELS
-                 "arrow_flight")
-endif()
-
 if(ARROW_BUILD_BENCHMARKS)
   # Perf server for benchmarks
   set(PERF_PROTO_GENERATED_FILES "${CMAKE_CURRENT_BINARY_DIR}/perf.pb.cc"
