+++ cpp/src/arrow/flight/test_definitions.cc
@@ -21,10 +21,17 @@
 
 #include "arrow/array/array_base.h"
 #include "arrow/array/array_dict.h"
+#include "arrow/array/util.h"
 #include "arrow/flight/api.h"
 #include "arrow/flight/test_util.h"
+#include "arrow/table.h"
 #include "arrow/testing/generator.h"
+#include "arrow/util/config.h"
 #include "arrow/util/logging.h"
+
+#if defined(ARROW_CUDA)
+#include "arrow/gpu/cuda_api.h"
+#endif
 
 namespace arrow {
 namespace flight {
@@ -1007,6 +1014,218 @@ void IpcOptionsTest::TestDoExchangeServerWriteOptions() {
   ASSERT_OK(writer->DoneWriting());
   ASSERT_RAISES(Invalid, writer->Close());
 }
+
+//------------------------------------------------------------
+// Test CUDA memory in data plane methods
+
+#if defined(ARROW_CUDA)
+
+Status CheckBuffersOnDevice(const Array& array, const Device& device) {
+  if (array.num_fields() != 0) {
+    return Status::NotImplemented("Nested arrays");
+  }
+  for (const auto& buffer : array.data()->buffers) {
+    if (!buffer) continue;
+    if (!buffer->device()->Equals(device)) {
+      return Status::Invalid("Expected buffer on device: ", device.ToString(),
+                             ". Was allocated on device: ", buffer->device()->ToString());
+    }
+  }
+  return Status::OK();
+}
+
+// Copy a record batch to host memory.
+arrow::Result<std::shared_ptr<RecordBatch>> CopyBatchToHost(const RecordBatch& batch) {
+  auto mm = CPUDevice::Instance()->default_memory_manager();
+  ArrayVector arrays;
+  for (const auto& column : batch.columns()) {
+    std::shared_ptr<ArrayData> data = column->data()->Copy();
+    if (data->child_data.size() != 0) {
+      return Status::NotImplemented("Nested arrays");
+    }
+
+    for (size_t i = 0; i < data->buffers.size(); i++) {
+      const auto& buffer = data->buffers[i];
+      if (!buffer || buffer->is_cpu()) continue;
+      ARROW_ASSIGN_OR_RAISE(data->buffers[i], Buffer::Copy(buffer, mm));
+    }
+    arrays.push_back(MakeArray(data));
+  }
+  return RecordBatch::Make(batch.schema(), batch.num_rows(), std::move(arrays));
+}
+
+class CudaTestServer : public FlightServerBase {
+ public:
+  explicit CudaTestServer(std::shared_ptr<Device> device) : device_(std::move(device)) {}
+
+  Status DoGet(const ServerCallContext&, const Ticket&,
+               std::unique_ptr<FlightDataStream>* data_stream) override {
+    RecordBatchVector batches;
+    RETURN_NOT_OK(ExampleIntBatches(&batches));
+    ARROW_ASSIGN_OR_RAISE(auto batch_reader, RecordBatchReader::Make(batches));
+    *data_stream = std::unique_ptr<FlightDataStream>(new RecordBatchStream(batch_reader));
+    return Status::OK();
+  }
+
+  Status DoPut(const ServerCallContext&, std::unique_ptr<FlightMessageReader> reader,
+               std::unique_ptr<FlightMetadataWriter> writer) override {
+    RecordBatchVector batches;
+    RETURN_NOT_OK(reader->ReadAll(&batches));
+    for (const auto& batch : batches) {
+      for (const auto& column : batch->columns()) {
+        RETURN_NOT_OK(CheckBuffersOnDevice(*column, *device_));
+      }
+    }
+    return Status::OK();
+  }
+
+  Status DoExchange(const ServerCallContext& context,
+                    std::unique_ptr<FlightMessageReader> reader,
+                    std::unique_ptr<FlightMessageWriter> writer) override {
+    FlightStreamChunk chunk;
+    bool begun = false;
+    while (true) {
+      RETURN_NOT_OK(reader->Next(&chunk));
+      if (!chunk.data) break;
+      if (!begun) {
+        begun = true;
+        RETURN_NOT_OK(writer->Begin(chunk.data->schema()));
+      }
+      for (const auto& column : chunk.data->columns()) {
+        RETURN_NOT_OK(CheckBuffersOnDevice(*column, *device_));
+      }
+      RETURN_NOT_OK(writer->WriteRecordBatch(*chunk.data));
+    }
+    return Status::OK();
+  }
+
+ private:
+  std::shared_ptr<Device> device_;
+};
+
+// Store CUDA objects without exposing them in the public header
+class CudaDataTest::Impl {
+ public:
+  cuda::CudaDeviceManager* manager;
+  std::shared_ptr<cuda::CudaDevice> device;
+  std::shared_ptr<cuda::CudaContext> context;
+};
+
+void CudaDataTest::SetUp() {
+  ASSERT_OK_AND_ASSIGN(auto manager, cuda::CudaDeviceManager::Instance());
+  ASSERT_OK_AND_ASSIGN(auto device, manager->GetDevice(0));
+  ASSERT_OK_AND_ASSIGN(auto context, device->GetContext());
+  impl_.reset(new Impl());
+  impl_->manager = manager;
+  impl_->device = std::move(device);
+  impl_->context = std::move(context);
+
+  ASSERT_OK_AND_ASSIGN(auto location, Location::ForScheme(transport(), "localhost", 0));
+  ASSERT_OK(MakeServer<CudaTestServer>(
+      location, &server_, &client_,
+      [this](FlightServerOptions* options) {
+        options->memory_manager = impl_->device->default_memory_manager();
+        return Status::OK();
+      },
+      [](FlightClientOptions* options) { return Status::OK(); }, impl_->device));
+}
+void CudaDataTest::TearDown() {
+  ASSERT_OK(client_->Close());
+  ASSERT_OK(server_->Shutdown());
+}
+void CudaDataTest::TestDoGet() {
+  // Check that we can allocate the results of DoGet with a custom
+  // memory manager.
+  FlightCallOptions options;
+  options.memory_manager = impl_->device->default_memory_manager();
+
+  Ticket ticket{""};
+  std::unique_ptr<FlightStreamReader> stream;
+  ASSERT_OK(client_->DoGet(options, ticket, &stream));
+  std::shared_ptr<Table> table;
+  ASSERT_OK(stream->ReadAll(&table));
+
+  for (const auto& column : table->columns()) {
+    for (const auto& chunk : column->chunks()) {
+      ASSERT_OK(CheckBuffersOnDevice(*chunk, *impl_->device));
+    }
+  }
+}
+void CudaDataTest::TestDoPut() {
+  RecordBatchVector batches;
+  ASSERT_OK(ExampleIntBatches(&batches));
+
+  std::unique_ptr<FlightStreamWriter> writer;
+  std::unique_ptr<FlightMetadataReader> reader;
+  auto descriptor = FlightDescriptor::Path({""});
+  ASSERT_OK(client_->DoPut(descriptor, batches[0]->schema(), &writer, &reader));
+
+  ipc::DictionaryMemo memo;
+  for (const auto& batch : batches) {
+    ASSERT_OK_AND_ASSIGN(auto buffer,
+                         cuda::SerializeRecordBatch(*batch, impl_->context.get()));
+    ASSERT_OK_AND_ASSIGN(auto cuda_batch,
+                         cuda::ReadRecordBatch(batch->schema(), &memo, buffer));
+
+    for (const auto& column : cuda_batch->columns()) {
+      ASSERT_OK(CheckBuffersOnDevice(*column, *impl_->device));
+    }
+
+    ASSERT_OK(writer->WriteRecordBatch(*cuda_batch));
+  }
+  ASSERT_OK(writer->Close());
+}
+void CudaDataTest::TestDoExchange() {
+  FlightCallOptions options;
+  options.memory_manager = impl_->device->default_memory_manager();
+
+  RecordBatchVector batches;
+  ASSERT_OK(ExampleIntBatches(&batches));
+
+  std::unique_ptr<FlightStreamWriter> writer;
+  std::unique_ptr<FlightStreamReader> reader;
+  auto descriptor = FlightDescriptor::Path({""});
+  ASSERT_OK(client_->DoExchange(options, descriptor, &writer, &reader));
+  ASSERT_OK(writer->Begin(batches[0]->schema()));
+
+  ipc::DictionaryMemo write_memo;
+  ipc::DictionaryMemo read_memo;
+  for (const auto& batch : batches) {
+    ASSERT_OK_AND_ASSIGN(auto buffer,
+                         cuda::SerializeRecordBatch(*batch, impl_->context.get()));
+    ASSERT_OK_AND_ASSIGN(auto cuda_batch,
+                         cuda::ReadRecordBatch(batch->schema(), &write_memo, buffer));
+
+    for (const auto& column : cuda_batch->columns()) {
+      ASSERT_OK(CheckBuffersOnDevice(*column, *impl_->device));
+    }
+
+    ASSERT_OK(writer->WriteRecordBatch(*cuda_batch));
+
+    FlightStreamChunk chunk;
+    ASSERT_OK(reader->Next(&chunk));
+    for (const auto& column : chunk.data->columns()) {
+      ASSERT_OK(CheckBuffersOnDevice(*column, *impl_->device));
+    }
+
+    // Bounce record batch back to host memory
+    ASSERT_OK_AND_ASSIGN(auto host_batch, CopyBatchToHost(*chunk.data));
+    AssertBatchesEqual(*batch, *host_batch);
+  }
+  ASSERT_OK(writer->Close());
+}
+
+#else
+
+void CudaDataTest::SetUp() {}
+void CudaDataTest::TearDown() {}
+void CudaDataTest::TestDoGet() { GTEST_SKIP() << "Arrow was built without ARROW_CUDA"; }
+void CudaDataTest::TestDoPut() { GTEST_SKIP() << "Arrow was built without ARROW_CUDA"; }
+void CudaDataTest::TestDoExchange() {
+  GTEST_SKIP() << "Arrow was built without ARROW_CUDA";
+}
+
+#endif
 
 }  // namespace flight
 }  // namespace arrow
